Major Update Changes the project structure so user don't have to sign in to see the blogs but it they try to perform any crud operations they have to sign in

+++ src/Routes/SpecificBlog/SpecificBlog.jsx
@@ -3,20 +3,21 @@ import { Link, useParams, useNavigate } from "react-router-dom";
 import { useSelector } from "react-redux";
 import { selectDarkMode } from "../../Features/ToggleModeSlice";
 import axios from "axios";
-import { ToastContainer, toast } from "react-toastify";
-import "react-toastify/dist/ReactToastify.css";
+import { toast } from "react-toastify";
 import Comments from "../Comments/Comments";
 import LikeButton from "../Likes/LikeButton";
 import Loader from "../../loader/Loader";
 import styles from "./SpecificBlog.module.css";
+import { handleAuthCheck } from "../../utils/AuthCheck";
 
 const SpecificBlog = () => {
   const darkMode = useSelector(selectDarkMode);
   const [blog, setBlog] = useState({});
-  const [loading, setLoading] = useState(true); // Add loading state
+  const [loading, setLoading] = useState(true);
   const { postId } = useParams();
-
   const navigate = useNavigate();
+  const JWT_TOKEN = localStorage.getItem("token");
+
   const fetchPost = async () => {
     try {
       const apiUrl = import.meta.env.VITE_REACT_APP_API_URL;
@@ -25,9 +26,8 @@ const SpecificBlog = () => {
       setBlog(res.post);
     } catch (error) {
       console.error("Error fetching post:", error);
-      // Handle error fetching post (e.g., redirect to an error page)
     } finally {
-      setLoading(false); // Set loading to false regardless of success or failure
+      setLoading(false);
     }
   };
 
@@ -36,10 +36,11 @@ const SpecificBlog = () => {
   }, [postId]);
 
   const deletePost = async () => {
+    if (!handleAuthCheck(JWT_TOKEN)) return;
+
     try {
       const apiUrl = import.meta.env.VITE_REACT_APP_API_URL;
       const url = `${apiUrl}/posts/${postId}`;
-      const JWT_TOKEN = localStorage.getItem("token");
       await axios.delete(url, {
         headers: {
           Authorization: `Bearer ${JWT_TOKEN}`,
@@ -59,7 +60,6 @@ const SpecificBlog = () => {
     }
   };
 
-  // Show loading indicator while data is being fetched
   if (loading) {
     return (
       <div
@@ -71,7 +71,6 @@ const SpecificBlog = () => {
     );
   }
 
-  // Render the content once data is fetched
   return (
     <div
       className={`${styles.container} ${darkMode ? styles.dark : styles.light}`}
@@ -82,19 +81,35 @@ const SpecificBlog = () => {
       <p className={styles.description}>{blog.description}</p>
 
       <div className={styles.buttons}>
-        <Link className={styles.link} to={`/update-post/${postId}`}>
+        <Link
+          className={styles.link}
+          to={`/update-post/${postId}`}
+          onClick={(e) => {
+            if (!handleAuthCheck(JWT_TOKEN)) e.preventDefault();
+          }}
+        >
           Update Post
         </Link>
         <button className={styles.btn} onClick={deletePost}>
           Delete Post
         </button>
       </div>
-      <LikeButton postId={postId} />
+      <LikeButton
+        postId={postId}
+        onLikeDislike={() => {
+          if (!handleAuthCheck()) return;
+          // Like or dislike logic here
+        }}
+      />
       <div className={styles.comment_section}>
-        <Comments postId={postId} />
+        <Comments
+          postId={postId}
+          onCommentSubmit={() => {
+            if (!handleAuthCheck()) return;
+            // Comment submit logic here
+          }}
+        />
       </div>
-      {/* Toast container for notifications */}
-      <ToastContainer />
     </div>
   );
 };
